+++ backend/src/main/java/study/memoravel/aop/MemberContext.java
@@ -3,9 +3,9 @@ package study.memoravel.aop;
 import study.memoravel.exception.jwt.NoJwtException;
 
 public class MemberContext {
-    public static final ThreadLocal<Integer> MEMBER_CONTEXT = new ThreadLocal<>();
+    public static final ThreadLocal<Long> MEMBER_CONTEXT = new ThreadLocal<>();
 
-    public static int getCurrentMemberId() {
+    public static long getCurrentMemberId() {
         if (MEMBER_CONTEXT.get() != null) {
             return MEMBER_CONTEXT.get();
         }
